ref: move HTTPURLResponse header extension to own file
The case-insensitive header lookup extension isn't specific to rate
limits, so move it out of DefaultRateLimits.swift into
HTTPURLResponse+Sentry.swift. Replace the awkward static ObjC wrapper
with an instance method.
